services/chat_backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Settings
pydantic-settings==2.12.0

# Database
psycopg2-binary==2.9.10

# HTTP Client (для Ollama)
httpx==0.28.1

# Logging
python-json-logger==2.0.7
